Move LowLevelDatabaseException to Bolt\Exception
